Restructured scripts
